chore(gRPC): move `Merge` trait and implementations to `iota-grpc-server`
