+++ crates/iota-grpc-server/src/ledger_service/get_objects.rs
@@ -6,7 +6,6 @@ use futures::Stream;
 use iota_grpc_types::{
     field::{FieldMaskTree, FieldMaskUtil},
     google::rpc::bad_request::FieldViolation,
-    merge::Merge,
     v0::{
         error_reason::ErrorReason,
         ledger_service::{GetObjectsRequest, GetObjectsResponse, ObjectResult, object_result},
@@ -20,6 +19,7 @@ use prost_types::FieldMask;
 use crate::{
     constants::validate_max_message_size,
     error::{ObjectNotFoundError, RpcError},
+    merge::Merge,
     types::{GrpcReader, ObjectsStreamResult},
 };
