+++ crates/iota-grpc-server/src/lib.rs
@@ -9,6 +9,7 @@ pub mod constants;
 mod error;
 pub mod event_filter;
 pub mod ledger_service;
+pub mod merge;
 pub mod response;
 pub mod server;
 pub mod transaction_execution_service;
