+++ crates/iota-grpc-server/src/transaction_execution_service/mod.rs
@@ -10,7 +10,6 @@ use std::sync::Arc;
 use iota_grpc_types::{
     field::FieldMaskTree,
     google::rpc::bad_request::FieldViolation,
-    merge::Merge,
     v0::{
         error_reason::ErrorReason,
         transaction::ExecutedTransaction,
@@ -27,7 +26,7 @@ use iota_types::{
 use tonic::{Request, Response};
 pub use transaction::{CommandResultsReadSource, TransactionReadSource};
 
-use crate::{error::RpcError, types::GrpcReader};
+use crate::{error::RpcError, merge::Merge, types::GrpcReader};
 
 pub const EXECUTE_TRANSACTION_READ_MASK_DEFAULT: &str = "transaction.effects";
