+++ crates/iota-grpc-types/src/proto/iota/grpc/v0/checkpoint.rs
@@ -5,56 +5,10 @@
 include!("../../../generated/iota.grpc.v0.checkpoint.rs");
 include!("../../../generated/iota.grpc.v0.checkpoint.field_info.rs");
 
-use crate::{field::FieldMaskTree, merge::Merge, proto::TryFromProtoError, v0::bcs::BcsData};
+use crate::{proto::TryFromProtoError, v0::bcs::BcsData};
 
 // CheckpointSummary
 //
-
-impl TryFrom<iota_sdk_types::CheckpointSummary> for CheckpointSummary {
-    type Error = Box<dyn std::error::Error>;
-
-    fn try_from(summary: iota_sdk_types::CheckpointSummary) -> Result<Self, Self::Error> {
-        Self::merge_from(summary, &FieldMaskTree::new_wildcard())
-    }
-}
-
-impl Merge<iota_sdk_types::CheckpointSummary> for CheckpointSummary {
-    fn merge(
-        &mut self,
-        source: iota_sdk_types::CheckpointSummary,
-        mask: &FieldMaskTree,
-    ) -> Result<(), Box<dyn std::error::Error>> {
-        if mask.contains(Self::BCS_FIELD.name) {
-            self.bcs = BcsData::serialize(&source).ok();
-        }
-
-        if mask.contains(Self::DIGEST_FIELD.name) {
-            self.digest = Some(source.digest().into());
-        }
-
-        Ok(())
-    }
-}
-
-impl Merge<&CheckpointSummary> for CheckpointSummary {
-    fn merge(
-        &mut self,
-        source: &CheckpointSummary,
-        mask: &FieldMaskTree,
-    ) -> Result<(), Box<dyn std::error::Error>> {
-        let CheckpointSummary { bcs, digest } = source;
-
-        if mask.contains(Self::DIGEST_FIELD.name) {
-            self.digest = digest.clone();
-        }
-
-        if mask.contains(Self::BCS_FIELD.name) {
-            self.bcs = bcs.clone();
-        }
-
-        Ok(())
-    }
-}
 
 impl TryFrom<&CheckpointSummary> for iota_sdk_types::CheckpointSummary {
     type Error = TryFromProtoError;
@@ -80,53 +34,6 @@ impl CheckpointSummary {
 // CheckpointContents
 //
 
-impl TryFrom<iota_sdk_types::CheckpointContents> for CheckpointContents {
-    type Error = Box<dyn std::error::Error>;
-
-    fn try_from(value: iota_sdk_types::CheckpointContents) -> Result<Self, Self::Error> {
-        Self::merge_from(value, &FieldMaskTree::new_wildcard())
-    }
-}
-
-impl Merge<iota_sdk_types::CheckpointContents> for CheckpointContents {
-    fn merge(
-        &mut self,
-        source: iota_sdk_types::CheckpointContents,
-        mask: &FieldMaskTree,
-    ) -> Result<(), Box<dyn std::error::Error>> {
-        if mask.contains(Self::BCS_FIELD.name) {
-            // TODO: add version
-            self.bcs = BcsData::serialize(&source).ok();
-        }
-
-        if mask.contains(Self::DIGEST_FIELD.name) {
-            self.digest = Some(source.digest().into());
-        }
-
-        Ok(())
-    }
-}
-
-impl Merge<&CheckpointContents> for CheckpointContents {
-    fn merge(
-        &mut self,
-        source: &CheckpointContents,
-        mask: &FieldMaskTree,
-    ) -> Result<(), Box<dyn std::error::Error>> {
-        let CheckpointContents { bcs, digest } = source;
-
-        if mask.contains(Self::BCS_FIELD.name) {
-            self.bcs = bcs.clone();
-        }
-
-        if mask.contains(Self::DIGEST_FIELD.name) {
-            self.digest = digest.clone();
-        }
-
-        Ok(())
-    }
-}
-
 impl TryFrom<&CheckpointContents> for iota_sdk_types::CheckpointContents {
     type Error = TryFromProtoError;
 
@@ -150,87 +57,6 @@ impl CheckpointContents {
 
 // Checkpoint
 //
-
-impl Merge<&iota_sdk_types::CheckpointSummary> for Checkpoint {
-    fn merge(
-        &mut self,
-        source: &iota_sdk_types::CheckpointSummary,
-        mask: &FieldMaskTree,
-    ) -> Result<(), Box<dyn std::error::Error>> {
-        if let Some(submask) = mask.subtree(Self::SUMMARY_FIELD.name) {
-            self.summary = Some(CheckpointSummary::merge_from(source.clone(), &submask)?);
-        }
-
-        Ok(())
-    }
-}
-
-impl Merge<iota_sdk_types::ValidatorAggregatedSignature> for Checkpoint {
-    fn merge(
-        &mut self,
-        source: iota_sdk_types::ValidatorAggregatedSignature,
-        mask: &FieldMaskTree,
-    ) -> Result<(), Box<dyn std::error::Error>> {
-        if mask.contains(Self::SIGNATURE_FIELD.name) {
-            self.signature = Some(source.into());
-        }
-
-        Ok(())
-    }
-}
-
-impl Merge<iota_sdk_types::CheckpointContents> for Checkpoint {
-    fn merge(
-        &mut self,
-        source: iota_sdk_types::CheckpointContents,
-        mask: &FieldMaskTree,
-    ) -> Result<(), Box<dyn std::error::Error>> {
-        if let Some(submask) = mask.subtree(Self::CONTENTS_FIELD.name) {
-            self.contents = Some(CheckpointContents::merge_from(source, &submask)?);
-        }
-
-        Ok(())
-    }
-}
-
-impl Merge<&Checkpoint> for Checkpoint {
-    fn merge(
-        &mut self,
-        source: &Checkpoint,
-        mask: &FieldMaskTree,
-    ) -> Result<(), Box<dyn std::error::Error>> {
-        let Checkpoint {
-            sequence_number,
-            summary,
-            signature,
-            contents,
-        } = source;
-
-        if mask.contains(Self::SEQUENCE_NUMBER_FIELD.name) {
-            self.sequence_number = *sequence_number;
-        }
-
-        if let Some(submask) = mask.subtree(Self::SUMMARY_FIELD.name) {
-            self.summary = summary
-                .as_ref()
-                .map(|summary| CheckpointSummary::merge_from(summary, &submask))
-                .transpose()?;
-        }
-
-        if mask.contains(Self::SIGNATURE_FIELD.name) {
-            self.signature = signature.clone();
-        }
-
-        if let Some(submask) = mask.subtree(Self::CONTENTS_FIELD.name) {
-            self.contents = contents
-                .as_ref()
-                .map(|contents| CheckpointContents::merge_from(contents, &submask))
-                .transpose()?;
-        }
-
-        Ok(())
-    }
-}
 
 impl Checkpoint {
     /// Deserialize checkpoint summary.
