+++ crates/iota-grpc-types/src/proto/iota/grpc/v0/signatures.rs
@@ -5,7 +5,7 @@
 include!("../../../generated/iota.grpc.v0.signatures.rs");
 include!("../../../generated/iota.grpc.v0.signatures.field_info.rs");
 
-use crate::{field::FieldMaskTree, merge::Merge, proto::TryFromProtoError, v0::bcs::BcsData};
+use crate::{proto::TryFromProtoError, v0::bcs::BcsData};
 
 // ValidatorAggregatedSignature
 //
@@ -37,56 +37,9 @@ impl TryFrom<iota_sdk_types::UserSignature> for UserSignature {
     type Error = Box<dyn std::error::Error>;
 
     fn try_from(value: iota_sdk_types::UserSignature) -> Result<Self, Self::Error> {
-        Self::merge_from(value, &FieldMaskTree::new_wildcard())
-    }
-}
-
-impl Merge<iota_types::signature::GenericSignature> for UserSignature {
-    fn merge(
-        &mut self,
-        source: iota_types::signature::GenericSignature,
-        mask: &FieldMaskTree,
-    ) -> Result<(), Box<dyn std::error::Error>> {
-        if !mask.contains(Self::BCS_FIELD.name) {
-            // No need to convert if no field is requested
-            return Ok(());
-        }
-
-        let sdk_signature: iota_sdk_types::UserSignature = source
-            .try_into()
-            .map_err(|e| format!("Failed to convert signature: {}", e))?;
-
-        Merge::merge(self, sdk_signature, mask)
-    }
-}
-
-impl Merge<iota_sdk_types::UserSignature> for UserSignature {
-    fn merge(
-        &mut self,
-        source: iota_sdk_types::UserSignature,
-        mask: &FieldMaskTree,
-    ) -> Result<(), Box<dyn std::error::Error>> {
-        if mask.contains(Self::BCS_FIELD.name) {
-            self.bcs = BcsData::serialize(&source).ok();
-        }
-
-        Ok(())
-    }
-}
-
-impl Merge<&UserSignature> for UserSignature {
-    fn merge(
-        &mut self,
-        source: &UserSignature,
-        mask: &FieldMaskTree,
-    ) -> Result<(), Box<dyn std::error::Error>> {
-        let UserSignature { bcs } = source;
-
-        if mask.contains(Self::BCS_FIELD.name) {
-            self.bcs = bcs.clone();
-        }
-
-        Ok(())
+        Ok(Self {
+            bcs: Some(BcsData::serialize(&value)?),
+        })
     }
 }
 
@@ -105,68 +58,6 @@ impl TryFrom<&UserSignature> for iota_sdk_types::UserSignature {
 
 // UserSignatures
 //
-
-impl Merge<iota_types::transaction::Transaction> for UserSignatures {
-    fn merge(
-        &mut self,
-        source: iota_types::transaction::Transaction,
-        mask: &FieldMaskTree,
-    ) -> Result<(), Box<dyn std::error::Error>> {
-        // Get signatures directly from transaction without converting the whole
-        // transaction
-        let tx_signatures = source.tx_signatures();
-
-        self.signatures = tx_signatures
-            .iter()
-            .map(|sig| {
-                // Convert iota_types signature to SDK signature, then merge
-                let sdk_sig: iota_sdk_types::UserSignature = sig
-                    .clone()
-                    .try_into()
-                    .map_err(|e| format!("Failed to convert signature: {e}"))?;
-                UserSignature::merge_from(sdk_sig, mask)
-            })
-            .collect::<Result<Vec<_>, _>>()?;
-
-        Ok(())
-    }
-}
-
-impl Merge<&iota_sdk_types::SignedTransaction> for UserSignatures {
-    fn merge(
-        &mut self,
-        source: &iota_sdk_types::SignedTransaction,
-        mask: &FieldMaskTree,
-    ) -> Result<(), Box<dyn std::error::Error>> {
-        if let Some(signatures_mask) = mask.subtree(Self::SIGNATURES_FIELD.name) {
-            self.signatures = source
-                .signatures
-                .iter()
-                .map(|sig| UserSignature::merge_from(sig.clone(), &signatures_mask))
-                .collect::<Result<Vec<_>, _>>()?;
-        }
-
-        Ok(())
-    }
-}
-
-impl Merge<&UserSignatures> for UserSignatures {
-    fn merge(
-        &mut self,
-        source: &UserSignatures,
-        mask: &FieldMaskTree,
-    ) -> Result<(), Box<dyn std::error::Error>> {
-        if let Some(signatures_mask) = mask.subtree(Self::SIGNATURES_FIELD.name) {
-            self.signatures = source
-                .signatures
-                .iter()
-                .map(|sig| UserSignature::merge_from(sig, &signatures_mask))
-                .collect::<Result<Vec<_>, _>>()?;
-        }
-
-        Ok(())
-    }
-}
 
 // TryFrom implementation for UserSignatures
 impl TryFrom<&UserSignatures> for Vec<iota_sdk_types::UserSignature> {
